token: encapsulate struct TokenList
The definition of struct TokenList can be opaque for better
encapsulation.
