+++ token/list/TokenList.h
@@ -1,34 +1,11 @@
 #ifndef TOKENLIST
 #define TOKENLIST
 
-#include <string.h>
 #include <stdbool.h>
 #include <inttypes.h>
 
 struct Token;
-
-struct TokenList {
-	//relative path of the source file
-	char* rel_path;
-
-	//all the tokens, including those
-	//already consumed
-	struct Token** tokens; //private
-	//how many pointers we can store
-	//with the currently allocated memory
-	uint16_t capacity; //private
-
-	//the index of the first token that was
-	//not already consumed
-	uint16_t index_head;
-
-	//the amount of tokens stored,
-	//even if they have already been consumed
-	uint16_t tokens_stored;
-
-	//token count
-	uint16_t tokensc; //private
-};
+struct TokenList;
 
 struct TokenList* makeTokenList();
 struct TokenList* makeTokenList2(char* filename);
@@ -55,6 +32,8 @@ int list_size(struct TokenList* list);
 char* list_code(struct TokenList* list);
 
 void list_print(struct TokenList* list);
+
+char* list_rel_path(struct TokenList* list);
 
 void freeTokenList(struct TokenList* list);
 void freeTokenListShallow(struct TokenList* list);
